collision

+++ Exercise1/Rover.cpp
@@ -10,17 +10,17 @@ Rover::Rover() {};
 
 Rover::Rover(MyObject obj) {
 	object = obj;
-	angle = 0;
-	speed = 0;
-	position[0] = 0;
-	position[1] = 0;
-	position[2] = 0;
+	angle = 0.0f;
+	speed = 0.0f;
+	position[0] = 0.0f;
+	position[1] = 0.0f;
+	position[2] = 0.0f;
 	updateDirection();
 }
 
 void Rover::updateDirection() {
-	float pi = 3.1415;
-	this->direction[0] = cos((pi / 180) * this->angle);
-	this->direction[2] = sin((pi / 180) * this->angle);
-	this->direction[1] = 0;
+	float pi = 3.1415f;
+	this->direction[0] = float(cos((pi / 180) * this->angle));
+	this->direction[2] = float(sin((pi / 180) * this->angle));
+	this->direction[1] = 0.0f;
 }
